refactor/#2: commons/lib 도메인 폴더 분리 및 fsd-instructure.md 추가
- commons/lib/error/ 폴더 신설: domainError, errorMessage, extractFieldErrorMessage 이동
- commons/lib/http/ 폴더 신설: apiClient, apiResponse, apiTypes 이동
- apps/route/ 폴더 신설: routeGuard를 commons에서 apps 레이어로 이동
- 15개 Route Handler의 apiResponse·withAuth import 경로 일괄 수정
- commons/lib/index.ts re-export 경로 갱신 (server-only 모듈 제외 유지)
- fsd-instructure.md 작성: 레이어 구조, 세그먼트 규칙, 네이밍 컨벤션 문서화
