+++ app/api/fridge/batch-usage/route.ts
@@ -1,7 +1,8 @@
 import { type NextRequest } from 'next/server';
 
-import { apiResponse } from '@/commons/lib/apiResponse';
-import { withAuth } from '@/commons/lib/routeGuard';
+import { withAuth } from '@/apps/route';
+
+import { apiResponse } from '@/commons/lib/http/apiResponse';
 
 /** GET /api/fridge/batch-usage?batchId= — 배치별 식단 사용량 합계 조회 */
 export const GET = withAuth(async (req: NextRequest, { supabase }) => {
