+++ app/api/fridge/batches/route.ts
@@ -1,8 +1,9 @@
 import { type NextRequest } from 'next/server';
 
+import { withAuth } from '@/apps/route';
+
 import { resolveDomainError } from '@/commons/lib';
-import { apiResponse } from '@/commons/lib/apiResponse';
-import { withAuth } from '@/commons/lib/routeGuard';
+import { apiResponse } from '@/commons/lib/http/apiResponse';
 import { type Json } from '@/commons/types';
 
 /** POST /api/fridge/batches — 기존 아이템에 배치 추가 */
