+++ app/api/fridge/preferences/route.ts
@@ -1,7 +1,8 @@
 import { type NextRequest } from 'next/server';
 
-import { apiResponse } from '@/commons/lib/apiResponse';
-import { withAuth } from '@/commons/lib/routeGuard';
+import { withAuth } from '@/apps/route';
+
+import { apiResponse } from '@/commons/lib/http/apiResponse';
 
 /** GET /api/fridge/preferences — 내 냉장고 표시 설정 조회 */
 export const GET = withAuth(async (_req: NextRequest, { userId, supabase }) => {
