+++ app/api/ingredients/stores/route.ts
@@ -2,8 +2,9 @@ import { type NextRequest } from 'next/server';
 
 import { uniq } from 'es-toolkit';
 
-import { apiResponse } from '@/commons/lib/apiResponse';
-import { withAuth } from '@/commons/lib/routeGuard';
+import { withAuth } from '@/apps/route';
+
+import { apiResponse } from '@/commons/lib/http/apiResponse';
 
 /** GET /api/ingredients/stores?householdId= — 구매처 목록 조회 (자동완성용) */
 export const GET = withAuth(async (req: NextRequest, { supabase }) => {
