+++ app/api/profile/route.ts
@@ -1,7 +1,8 @@
 import { type NextRequest } from 'next/server';
 
-import { apiResponse } from '@/commons/lib/apiResponse';
-import { withAuth } from '@/commons/lib/routeGuard';
+import { withAuth } from '@/apps/route';
+
+import { apiResponse } from '@/commons/lib/http/apiResponse';
 
 /** GET /api/profile — 내 프로필 조회 */
 export const GET = withAuth(async (_req: NextRequest, { userId, supabase }) => {
