+++ src/apps/route/index.ts
@@ -1,0 +1,1 @@
+export { withAuth, type AuthContext } from './routeGuard';
