+++ src/commons/lib/error/domainError.ts
