+++ src/commons/lib/index.ts
@@ -1,10 +1,10 @@
-export { apiClient } from './apiClient';
-export { type ApiResponse } from './apiTypes';
+export { apiClient } from './http/apiClient';
+export { type ApiResponse } from './http/apiTypes';
 export { cn } from './cn';
 export { createSafeContext } from './context';
-export { DOMAIN_ERROR_CODE, DOMAIN_ERROR_MESSAGE, resolveDomainError } from './domainError';
-export { ERROR_MSG } from './errorMessage';
-export { extractFieldErrorMessage } from './extractFieldErrorMessage';
+export { DOMAIN_ERROR_CODE, DOMAIN_ERROR_MESSAGE, resolveDomainError } from './error/domainError';
+export { ERROR_MSG } from './error/errorMessage';
+export { extractFieldErrorMessage } from './error/extractFieldErrorMessage';
 export {
   getCurrentPushSubscription,
   isNotificationSupported,
@@ -18,5 +18,5 @@ export {
 export { parseSafeNumericInput } from './numericInput';
 export { uuid } from './uuid';
 export { cva, type VariantProps } from 'class-variance-authority';
-// apiResponse, routeGuard, withAuth는 server-only (next/headers 의존)
-// Route Handler에서 직접 import: '@/commons/lib/apiResponse', '@/commons/lib/routeGuard'
+// apiResponse는 server-only (next/server 의존) → Route Handler에서 직접 import: '@/commons/lib/http/apiResponse'
+// withAuth, AuthContext → '@/apps/route'
